Implement 4: isEven
After racking my brain for a solution that didn't simply take one step (and
miss the chance for recursion), I looked online for algorithms.

After this, I'm going to be quick to check online for algorithms whenever the
function involves numbers. This isn't about constructing algorithms; it's
about implementing them.